% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_of_SL.R
\name{count_of_SL}
\alias{count_of_SL}
\title{A Count Function to process SLIPT predictions}
\usage{
count_of_SL(x, significance = "fdr", threshold = 0.05)
}
\arguments{
\item{x}{Data matrix. Synthetic Lethal predictions to process, typically the output of \code{\link[slipt]{detect_SL}}}

\item{significance}{String. Significance condition for p-values to use: none, raw, or adjusted with a valid method for p.adjust() Defaults to fdr (false discovery rate / BH).}

\item{threshold}{Numeric. Significance threshold (alpha, type I error rate) to cut-off (raw or adjusted) p-values. Defaults to 0.05.}
}
\description{
Count number of predicted synthetic lethal gene partners with synthetic lethal conditions and significance thresholds met.
}
\examples{
count_of_SL()

}
\keyword{gene}
\keyword{interaction,}
\keyword{lethal,}
\keyword{post-processing,}
\keyword{significance,}
\keyword{slice}
\keyword{subset,}
\keyword{synthetic}

