% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_SL.R
\name{detect_SL}
\alias{detect_SL}
\title{Prediction of Synthetic Lethality}
\usage{
detect_SL(query, datasetx)
}
\arguments{
\item{query}{string. Name of a gene to test all others against. Note this gene must be contained within the dataset, matching a column name of datasetx.}

\item{datasetx}{Dataset input to be formatted. Input data is: samples (rows) x genes (columns) as formatted by \code{\link[slipt]{prep_data_for_SL}}. Output data is a table of genes (rows) x test statistics and outputs (columns), in order of most significantly supported candidate interacting parnter, this tabular format is human-readable is handled by \code{\link[slipt]{count_of_SL}} or \code{\link[slipt]{table_of_SL}} for built-in summary statistics.}
}
\description{
Statistical detection of expression signals consistent with synthetic lethal interaction. Based on the Chi-Square test with built-in multiple testing corrections.
}
\examples{
#prepare data
 data <- c()
for(i in 1:100){
  data <- cbind(data, rnorm(1000))
}
rm(i)
rownames(data) <- paste("gene", 1:1000)
colnames(data) <- paste("sample", 1:100)
partitioned_data <- prep_data_for_SL(data, n = 3)

#run SLIPT
sl_table <- detect_SL("gene 1", partitioned_data)
dim(sl_table)
head(dim(sl_table))

sl_table <- detect_SL("gene 1", prep_data_for_SL(data))
dim(sl_table)
head(dim(sl_table))

}
\keyword{chi-square,}
\keyword{expression,}
\keyword{genetics,}
\keyword{genomics}
\keyword{lethal,}
\keyword{synthetic}
