% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_SL_latex.R
\name{export_SL_latex}
\alias{export_SL_latex}
\title{An Export Function to process SLIPT predictions}
\usage{
export_SL_latex(x)
}
\arguments{
\item{x}{Data matrix. Synthetic Lethal predictions to process, typically the output of \code{\link[slipt]{detect_SL}} or \code{\link[slipt]{table_of_SL}}.}
}
\description{
Exports the key synthetic lethal gene information in a LaTeX table. Can be done for all genes (output of \code{\link[slipt]{detect_SL}}) or with synthetic lethal conditions and significance thresholds met (output of \code{\link[slipt]{table_of_SL}}).
}
\examples{
#prepare data
 data <- c()
for(i in 1:100){
  data <- cbind(data, rnorm(1000))
}
rm(i)
rownames(data) <- paste("gene", 1:1000)
colnames(data) <- paste("sample", 1:100)
partitioned_data <- prep_data_for_SL(data, n = 3)

#run SLIPT
sl_table <- detect_SL("gene 1", partitioned_data)
dim(sl_table)
head(dim(sl_table))

sl_table <- detect_SL("gene 1", prep_data_for_SL(data))
dim(sl_table)
head(dim(sl_table))

#export top results to TeX
export_SL_latex(sl_table)

}
\keyword{LaTeX,}
\keyword{gene}
\keyword{interaction,}
\keyword{lethal,}
\keyword{post-processing,}
\keyword{significance,}
\keyword{slice,}
\keyword{subset,}
\keyword{synthetic}
\keyword{table}
\keyword{tex,}
