% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data_for_SL.R
\name{prep_data_for_SL}
\alias{prep_data_for_SL}
\title{Prepare Dataset for SLIPT (Synthetic Lethal Predition)}
\usage{
prep_data_for_SL(dataset, n = 3L, exclude = T)
}
\arguments{
\item{dataset}{Dataset input to be formatted. Input is genes (rows) x samples (columns). Output data is transposed: samples (rows) x genes (columns).}

\item{n}{number of quantiles to test. Must be an integer. Defaults to 3.}

\item{exclude}{Do you wish to exclude genes without quartile variation? Defaults to TRUE.}
}
\description{
Prepare data (e.g., gene expression data) for synthetic lethal prediction (typically count or continous data) by bining samples into categorical data for each gene to be tested.
}
\examples{
data <- c()
for(i in 1:100){
  data <- cbind(data, rnorm(1000))
}
rm(i)
rownames(data) <- paste("gene", 1:1000)
colnames(data) <- paste("sample", 1:100)
dim(data)

partitioned_data <- prep_data_for_SL(data, n = 3)
dim(partitioned_data)

}
\keyword{binning,}
\keyword{categorical,}
\keyword{lethal}
\keyword{quantile,}
\keyword{synthetic}
